requirements.txt
dj-stripe==2.5.0
Django==3.2.3
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-phonenumber-field==5.1.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-ajax-datatable~=4.2.1